Refactor role management and permissions
- Extended roles table migration to remove is_system and is_full_access columns.
- Retired positions and work schedules migrations as they are no longer needed.
- Created org_infos table migration for per-org-unit company information.
- Updated DatabaseSeeder to include OrgInfoSeeder for initial data population.
- Modified HrisRoleSeeder to adjust role metadata and permissions handling.
- Introduced OrgInfoSeeder to populate company info for specific org units.
- Added API routes for managing company information per org unit.
- Updated tests to reflect changes in role permissions and ghost user behavior.
- Implemented new tests for OrgInfo functionality and org unit scope switching.

+++ database/migrations/2026_06_01_000015_create_leave_types_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('code')->unique();
-            $table->decimal('max_days_per_year', 5, 2)->nullable();
+            $table->integer('max_days_per_year')->nullable();
             $table->boolean('requires_balance')->default(true);
             $table->boolean('requires_attachment')->default(false);
             $table->boolean('is_paid')->default(true);
